switched integer types to double within some components and also made jumpresetsystem

See merge request CompSci308_2018Spring/voogasalad_callussalad!312

+++ src/game_engine/components/DefaultNumberOfJumpsComponent.java
@@ -2,8 +2,8 @@ package game_engine.components;
 
 import game_engine.Component;
 
-public class DefaultNumberOfJumpsComponent extends Component<Integer> {
+public class DefaultNumberOfJumpsComponent extends Component<Double> {
 	public DefaultNumberOfJumpsComponent(String arg) {
-		super(Integer.parseInt(arg));
+		super(Double.parseDouble(arg));
 	}
 }
